Project Report- Vision section
Guys, Still we have to get 50% of the allocated marks for major project.
I have created a document and updated the vision section under
discussion. (Divided sections in the document are as the marking
scheme.)

Feel free to change anything as you wish

I know you guys don't use 'word' but this is the only option i got.